hopr_lib::api::chain

Struct WinningProbability

pub struct WinningProbability(/* private fields */);

Represents a ticket-winning probability.

It holds the modified IEEE-754 representation but behaves like a reduced precision float. It intentionally does not implement Ord or Eq, as it can be either approximately compared or lexicographically compared.

Implementations

impl WinningProbability

pub const ALWAYS: WinningProbability

100% winning probability

pub const EPSILON: f64 = 0.00000001

pub const NEVER: WinningProbability

0% winning probability.

pub fn as_luck(&self) -> u64

Converts winning probability to an unsigned integer (luck).

pub fn as_encoded(&self) -> [u8; 7]

Convenience function to convert to internal probability representation.

pub fn as_f64(&self) -> f64

Convert probability to a float.

pub fn try_from_f64(win_prob: f64) -> Result<WinningProbability, CoreTypesError>

Tries to get probability from a float.

pub fn approx_cmp(&self, other: &WinningProbability) -> Ordering

Performs approximate comparison up to Self::EPSILON.

pub fn approx_eq(&self, other: &WinningProbability) -> bool

Performs approximate equality comparison up to Self::EPSILON.

pub fn lex_cmp(&self, other: &WinningProbability) -> Ordering

Performs lexicographical comparison of the luck values.

pub fn lex_eq(&self, other: &WinningProbability) -> bool

Performs lexicographical equality comparison of the luck values.

pub fn min(&self, other: &WinningProbability) -> WinningProbability

Gets the minimum of two winning probabilities.

pub fn max(&self, other: &WinningProbability) -> WinningProbability

Gets the maximum of two winning probabilities.