hopr_lib::api

Module ct

Cover traffic related API traits.

Modules

traits

types

Enums

DestinationRouting

Routing information containing forward or return routing options.

Telemetry

Enum representing different types of telemetry data used by the CT mechanism.

TrafficGenerationError

Traits

MeasurableNeighbor

Measurable neighbor telemetry.

MeasurablePath

Measurable path telemetry.

NetworkGraphUpdate

A trait specifying the graph update functionality

NetworkGraphView

A trait specifying the graph traversal functionality

TrafficGeneration

A trait for types that can produce a stream of cover traffic routes.