hopr_lib::api

Module graph

Network graph related API traits.

Modules

costs

traits

types

Structs

EdgeCapacityUpdate

Update for the edge between src and dest.

Enums

EdgeTransportTelemetry

Enum representing different types of telemetry data used by the CT mechanism.

MeasurableEdge

Edge measurements accepted for an edge in the graph.

NetworkGraphError

Error observed during the measurements updating the graph edges.

Traits

CostFn

Definition of a fold like cost function usable with graph traversal trait.

EdgeLinkObservable

Trait for recording and querying transport-level link quality metrics for a transport link.

MeasurableNode

MeasurablePath

Measurable path segment telemetry.

MeasurablePeer

Measurable neighbor peer attributes.

NetworkGraphTraverse

A trait specifying the graph traversal functionality.

NetworkGraphUpdate

A trait for recording observed measurement updates to graph edges and nodes.

NetworkGraphView

A trait specifying read-only graph view functionality.

NetworkGraphWrite

A trait for mutating the graph topology.