hopr_lib::api::graph::traits

Trait EdgeObservableRead

pub trait EdgeObservableRead {
    type ImmediateMeasurement: EdgeLinkObservable + EdgeNetworkObservableRead + Send;
    type IntermediateMeasurement: EdgeLinkObservable + EdgeProtocolObservable + Send;

    // Required methods
    fn last_update(&self) -> Duration;
    fn immediate_qos(&self) -> Option<&Self::ImmediateMeasurement>;
    fn intermediate_qos(&self) -> Option<&Self::IntermediateMeasurement>;
    fn score(&self) -> f64;
}

Trait for reading aggregated quality-of-service observations from a graph edge.

Required Associated Types

type ImmediateMeasurement: EdgeLinkObservable + EdgeNetworkObservableRead + Send

Measurement type for direct (1-hop) probes, including network connectivity info.

type IntermediateMeasurement: EdgeLinkObservable + EdgeProtocolObservable + Send

Measurement type for relayed probes through an intermediate, including channel capacity.

Required Methods

fn last_update(&self) -> Duration

The timestamp of the last update.

fn immediate_qos(&self) -> Option<&Self::ImmediateMeasurement>

Transport level measurements between this node and any other node in the network.

fn intermediate_qos(&self) -> Option<&Self::IntermediateMeasurement>

Transport level measurements performed in a transparent mode using looping measurements.

fn score(&self) -> f64

A value scoring the observed peer.

It is from the [0.0, 1.0] range. The higher the value, the better the score.