hopr_lib::api::graph

Module types

Structs

EdgeCapacityUpdate

Update for the edge between src and dest.

Enums

EdgeTransportTelemetry

Enum representing different types of telemetry data used by the CT mechanism.

MeasurableEdge

Edge measurements accepted for an edge in the graph.

NetworkGraphError

Error observed during the measurements updating the graph edges.

Traits

MeasurableNode

MeasurablePath

Measurable path segment telemetry.

MeasurablePeer

Measurable neighbor peer attributes.