hopr_lib::api::network

Module types

Structs

BoundedVec

Wrapper for Vec that has an explicit upper bound on the number of elements. The Structure remains heap-allocated to avoid blowing up the size of types where it is used.

HealthIter

An iterator over the variants of Health

PeerId

Identifier of a peer of the network.

Enums

Health

Network health represented with colors, where green is the best and red is the worst possible observed network quality.

NetworkEvent

Events generated by the network for external consumption.

Type Aliases

PathId

Identifier for a path traversed using an allowed DestinationRouting over the network.