hopr_lib::api

Crate types

Contains all HOPR-specific or related Rust types.

The individual types are feature-gated.

Modules

chain

This crate contains various on-chain related modules and types.

crypto

This Rust crate contains implementation of common cryptographic types.

crypto_random

This Rust crate contains implementation of common random number generation functions. All functions and types from this crate supply cryptographically secure random numbers.

internal

This crate contains all types that are specific to the HOPR protocol. As opposed to hopr-primitive-types which contains more generic types (not necessarily specific only to HOPR).

primitive

This crate contains basic types used throughout the entire HOPR codebase. Types from this crate are not necessarily specific only to HOPR.