hopr_lib::api::types::chain::exports::alloy::signers::k256::elliptic_curve::scalar

Struct NonZeroScalar

pub struct NonZeroScalar<C>
where
    C: CurveArithmetic,
{ /* private fields */ }

Non-zero scalar type.

This type ensures that its value is not zero, ala core::num::NonZero*. To do this, the generic S type must impl both Default and ConstantTimeEq, with the requirement that S::default() returns 0.

In the context of ECC, it’s useful for ensuring that scalar multiplication cannot result in the point at infinity.

Implementations

impl<C> NonZeroScalar<C>
where C: CurveArithmetic,

pub fn random(rng: &mut impl CryptoRngCore) -> NonZeroScalar<C>

Generate a random NonZeroScalar.

pub fn new(scalar: <C as CurveArithmetic>::Scalar) -> CtOption<NonZeroScalar<C>>

Create a NonZeroScalar from a scalar.

pub fn from_repr( repr: GenericArray<u8, <C as Curve>::FieldBytesSize>, ) -> CtOption<NonZeroScalar<C>>

Decode a NonZeroScalar from a big endian-serialized field element.

pub fn from_uint(uint: <C as Curve>::Uint) -> CtOption<NonZeroScalar<C>>

Create a NonZeroScalar from a C::Uint.