hopr_lib
Module constants

Various public constants.

Constants

APP_VERSION

Application version as presented externally using the heartbeat mechanism

APP_VERSION_COERCED

Application version coerced into the major.minor.patch form

MIN_NATIVE_BALANCE

Minimum native token balance to start the node.

ONBOARDING_INFORMATION_INTERVAL

SUGGESTED_NATIVE_BALANCE

Balance that is suggested for funding the node the first time in order to make it start.