hopr_lib

Function chain_events_to_transport_events

pub async fn chain_events_to_transport_events<StreamIn, Db>(
    event_stream: StreamIn,
    me_onchain: Address,
    db: Db,
    multi_strategy: Arc<MultiStrategy>,
    channel_graph: Arc<RwLock<ChannelGraph>>,
    indexer_action_tracker: Arc<IndexerActionTracker>,
) -> impl Stream<Item = PeerDiscovery> + Send + 'static
where
    Db: HoprDbAllOperations + Clone + Send + Sync + Debug + 'static,
    StreamIn: Stream<Item = SignificantChainEvent> + Send + 'static,

Creates a pipeline that chains the indexer-generated data, processes them into the individual components, and creates a filtered output stream that is fed into the transport layer swarm.

event_stream - represents the events generated by the indexer. If the Indexer is not synced, it will not generate any events.

preloading_event_stream - a stream used by the components to preload the data from the objects (db, channel graph…)