hopr_lib

Module prelude

Export of relevant types for easier integration.

Structs

Address

Represents an Ethereum address

ChainKeypair

Represents a keypair consisting of a secp256k1 private and public key

ConnectedUdpStream

Mimics TCP-like stream functionality on a UDP socket by restricting it to a single counterparty and implements [tokio::io::AsyncRead] and [tokio::io::AsyncWrite]. The instance must always be constructed using a UdpStreamBuilder.

HoprKeys

OffchainKeypair

Represents a keypair consisting of an Ed25519 private and public key

OffchainPublicKey

Represents an Ed25519 public key.

Enums

ForeignDataMode

Defines what happens when data from another SocketAddr arrives into the ConnectedUdpStream (other than the one that is considered a counterparty for that instance).

UdpStreamParallelism

Determines how many parallel readers or writer sockets should be bound in ConnectedUdpStream.

Type Aliases

Hash

Represents an Ethereum 256-bit hash value.