hopr_lib

Crate session

HoprSession object providing the session functionality over the HOPR transport

The session proxies the user interactions with the transport to hide the advanced interactions and functionality.

The SessionManager allows for automatic management of sessions via the Start protocol.

This crate implements RFC-0007.

Modules

errors

Structs

AckSnapshot

Snapshot of acknowledgement metrics.

AtomicSurbFlowEstimator

An implementation of [SurbFlowEstimator] that tracks the number of produced and consumed SURBs via two AtomicU64s.

ByteCapabilities

Wrapper for Capabilities that makes conversion to/from u8 possible.

FrameBufferSnapshot

Snapshot of frame buffer metrics.

HoprSession

Represents the Session protocol socket over HOPR.

HoprSessionConfig

Configures the Session protocol socket over HOPR.

IncomingSession

Wrapper for incoming HoprSession along with other information extracted from the Start protocol during the session establishment.

SessionClientConfig

Configuration for the session.

SessionId

Unique ID of a specific Session in a certain direction.

SessionLifetimeSnapshot

Snapshot of session lifetime metrics.

SessionManager

Manages lifecycles of Sessions.

SessionManagerConfig

Configuration for the SessionManager.

SessionStatsSnapshot

Complete snapshot of all session metrics at a point in time.

SessionTelemetry

Internal metrics tracker for a session.

SurbBalancerConfig

Configuration for the SurbBalancer.

SurbSnapshot

Snapshot of SURB (Single Use Reply Block) metrics.

TransportSnapshot

Snapshot of transport-level data metrics.

Enums

AcknowledgementMode

Indicates the acknowledgement mode of a stateful Session socket.

Capability

Individual capabilities of a Session.

DestinationRouting

Routing information containing forward or return routing options.

DispatchResult

Indicates the result of processing a message.

IpOrHost

Implements a host name with port. This could be either a DNS name with port or an IP address with port represented by std::net::SocketAddr.

IpProtocol

Lists some of the IP protocols.

ResolvedTransportRouting

Contains the resolved routing information for the packet.

RoutingOptions

Represents routing options in a mixnet with a maximum number of hops.

SealedHost

Contains optionally encrypted IpOrHost.

SessionAckMode

The acknowledgement mode configured for the session.

SessionLifecycleState

The lifecycle state of a session from the perspective of metrics.

SessionTarget

Defines what should happen with the data at the recipient where the data from the established session are supposed to be forwarded to some target.

SurbMatcher

Allows finding saved SURBs based on different criteria.

Constants

MIN_BALANCER_SAMPLING_INTERVAL

Smallest possible interval for balancer sampling.

MIN_SURB_BUFFER_DURATION

Minimum time the SURB buffer must endure if no SURBs are being produced.

SESSION_MTU

Number of bytes that can be sent in a single Session protocol payload.

SURB_SIZE

Size of the HOPR SURB in bytes.

Functions

transfer_session

Convenience function to copy data in both directions between a Session and arbitrary async IO stream. This function is only available with Tokio and will panic with other runtimes.

Type Aliases

Capabilities

Set of Session capabilities.

ServiceId

Describes a node service target. These are specialized SessionTargets that are local to the Exit node and have different purposes, such as Cover Traffic.