Crate hopr_lib

HOPR library creating a unified Hopr object that can be used on its own, as well as integrated into other systems and libraries.

Re-exports

pub use core_strategy::Strategy::AutoRedeeming;

Modules

config

constants

errors

rlp

Structs

AccountEntry

Represents a node announcement entry on the block chain. This contains node’s public key and optional announcement information (multiaddress, block number).

AcknowledgedTicket

Contains acknowledgment information and the respective ticket

Acknowledgement

Represents packet acknowledgement

Address

Represents an Ethereum address

AnnouncementData

Structure containing data used for on-chain announcement. That is the decapsulated multiaddress (with the /p2p/{peer_id} suffix removed) and optional KeyBinding (announcement can be done with key bindings or without)

ApplicationData

Represents the received decrypted packet carrying the application-layer data.

Balance

Represents balance of some coin or token.

ChainNetwork

Holds all information about the protocol network to be used by the client

ChannelEntry

Overall description of a channel

CloseChannelResult

EthereumChallenge

Represents and Ethereum challenge.

HalfKeyChallenge

Represents a challenge for the half-key in Proof of Relay. Half-key challenge is equivalent to a secp256k1 curve point. Therefore, HalfKeyChallenge can be obtained from a HalfKey.

Hopr

HostConfig

KeyBinding

Holds the signed binding of the chain key and the packet key. The signature This is used to attest on-chain that node owns the corresponding packet key and links it with the chain key.

Multiaddr

Representation of a Multiaddr.

NoSumSMA

Basic implementation of Simple Moving Average (SMA). The maximum window size is bound by 2^32 - 1. Useful mainly for floating-point types, as it does not accumulate floating point error with each sample. Requires O(N) of memory and O(N) for average computation, N being window size. The divisor argument D is used only for such types T that do not implement From<u32> (such as Duration,…).

OpenChannelResult

ProtocolsConfig

SingleSumSMA

Basic implementation of Simple Moving Average (SMA). The maximum window size is bound by 2^32 - 1. Useful mainly for integer types, as it does accumulate floating point error with each sample. Requires O(N) of memory and O(1) for average computation, N being window size. The divisor argument D is used only for such types T that do not implement From<u32> (such as Duration,…).

Snapshot

Represents a snapshot in the blockchain

TagBloomFilter

Bloom filter for packet tags to detect packet replays. In addition, this structure also holds the number of items in the filter to determine if the filter needs to be refreshed. Once this happens, packet replays of past packets might be possible.

Ticket

Contains the overall description of a ticket with a signature

TicketStatistics

UnacknowledgedTicket

Wrapper for an unacknowledged ticket

Enums

AccountType

Type of the node account.

AcknowledgedTicketStatus

Status of the acknowledged ticket.

BalanceType

Represents a type of the balance: native or HOPR tokens.

ChannelChange

Enumerates possible changes on a channel entry update

ChannelDirection

Describes a direction of node’s own channel. The direction of a channel that is not own is undefined.

ChannelStatus

Describes status of a channel

CoreEthereumActionsError

CoreTypesError

GeneralError

Listing of some general re-usable errors

Health

HoprLoopComponents

Enum differentiator for loop component futures.

HoprTransportError

HostType

PendingAcknowledgement

Contains either unacknowledged ticket if we’re waiting for the acknowledgement as a relayer or information if we wait for the acknowledgement as a sender.

ProtocolError

State

Strategy

Enumerates all possible strategies with their respective configurations.

TransportOutput

Object representing different types of output from the transport layer

Constants

DEFAULT_APPLICATION_TAG

Represent a default application tag if none is specified in send_packet.

INTERMEDIATE_HOPS

Number of intermediate hops: 3 relayers and 1 destination

PAYLOAD_SIZE

Maximum size of the packet payload

PEER_METADATA_PROTOCOL_VERSION

Name of the metadata key holding the protocol version

TICKET_WIN_PROB

Fixed ticket winning probability

Traits

AutoBinarySerializable

Type implementing this trait has automatic binary serialization/deserialization capability using the default binary format, which is currently bincode.

BinarySerializable

A type that can be serialized and deserialized to a binary form. Implementing this trait automatically implements ToHex trait which then uses the serialize method.

Keypair

Represents a generic key pair The keypair contains a private key and public key. Must be comparable in constant time and zeroized on drop.

PeerAddressResolver

Trait for linking and resolving the corresponding OffchainPublicKey and on-chain Address.

SMA

Simple Moving Average trait. The second-most useful filter type, bested only by coffee filters.

ToHex

A generic type that can be converted to a hexadecimal string.

UnitaryFloatOps

Allows type to be multiplied and divided by a float in range [0.0, 1.0].

Functions

build_components

Main builder of the hopr lib components

f64_to_win_prob

Encodes [0.0f64, 1.0f64] to [0x00000000000000, 0xffffffffffffff]

generate_channel_id

looks_like_domain

to_chain_events_refresh_process

win_prob_to_f64

Decodes [0x00000000000000, 0xffffffffffffff] to [0.0f64, 1.0f64]

Type Aliases

EncodedWinProb

Tag

Tags are currently 16-bit unsigned integers

U256