hopr_lib
Trait BytesRepresentable

pub trait BytesRepresentable<E = GeneralError>: AsRef<[u8]> + for<'a> TryFrom<&'a [u8], Error = E> {
    const SIZE: usize;

    // Provided method
    fn into_boxed(self) -> Box<[u8]> { ... }
}

Represents a type that is already internally represented by a fixed size byte array, and therefore requires no memory allocation to represent the type in binary encoding.

This is a strict subset of BytesEncodable, see its documentation for details.

Required Associated Constants

const SIZE: usize

Size of the encoded byte array.

Provided Methods

fn into_boxed(self) -> Box<[u8]>

Convenience function to copy this type’s binary representation into a Box.

Object Safety

This trait is not object safe.