hopr_lib

Trait ToHex

pub trait ToHex {
    // Required methods
    fn to_hex(&self) -> String;
    fn from_hex(str: &str) -> Result<Self, GeneralError>
       where Self: Sized;
}

A generic type that can be converted to a hexadecimal string.

Implementors of this trait should automatically take care of the optional 0x prefix.

Required Methods

fn to_hex(&self) -> String

Hexadecimal representation of this type.

fn from_hex(str: &str) -> Result<Self, GeneralError>
where Self: Sized,

Tries to parse the type from the hexadecimal representation.